+++ src/Certes/Acme/Resource/Challenge.cs
@@ -1,5 +1,6 @@
 ﻿using Newtonsoft.Json;
 using System;
+using System.Collections.Generic;
 
 namespace Certes.Acme.Resource
 {
@@ -53,6 +54,16 @@ namespace Certes.Acme.Resource
         /// </value>
         [JsonProperty("error")]
         public AcmeError Error { get; set; }
+
+        /// <summary>
+        /// Gets or sets the errors.
+        /// </summary>
+        /// <value>
+        /// The errors.
+        /// </value>
+        [JsonProperty("errors")]
+        [Obsolete("Use Challenge.Error instead.")]
+        public IList<object> Errors { get; set; }
 
         /// <summary>
         /// Gets or sets the token.